third_party/WebKit/LayoutTests/accessibility/contenteditable-caret-position.html

<!DOCTYPE html>
<html>
  <head>
    <script src="../resources/js-test.js"></script>
  </head>
  <body>

    <div id="contenteditable-textbox" role="textbox" contenteditable="true">
      <div id="contenteditable-line1">Line 1</div>
      <textarea id="contenteditable-line2" rows="1" cols="40">Line 2</textarea>
    </div>

    <div id="contenteditable-div" contenteditable>
      <p id="paragraph1">Line 1<br>Line 2</p>
      <p id="paragraph2">Line 3</p>
    </div>

    <script>
        description("This tests that caret position is reported correctly for contenteditable elements.");

        if (window.accessibilityController) {
            var selection = window.getSelection();
            var selectionRange = document.createRange();

            var textbox = document.getElementById("contenteditable-textbox");
            textbox.focus();
            var textboxAccessible =
                accessibilityController.accessibleElementById(
                'contenteditable-textbox');

            shouldBeZero("textboxAccessible.selectionStart");
            shouldBeZero("textboxAccessible.selectionEnd");

            var line1 = document.getElementById("contenteditable-line1");
            selectionRange.setStart(line1.firstChild, 1);
            selectionRange.setEnd(line1.firstChild, 1);
            selection.removeAllRanges();
            selection.addRange(selectionRange);
            shouldBeEqualToNumber("textboxAccessible.selectionStart", 1);
            shouldBeEqualToNumber("textboxAccessible.selectionEnd", 1);

            var line2 = document.getElementById("contenteditable-line2");
            line2.focus();
            // The selection should have been removed from the line1 div.
            shouldBeZero("textboxAccessible.selectionStart");
            shouldBeZero("textboxAccessible.selectionEnd");

            var line2Accessible = accessibilityController.focusedElement;
            shouldBeZero("line2Accessible.selectionStart");
            shouldBeZero("line2Accessible.selectionEnd");
            line2.setSelectionRange(3, 3);
            shouldBeEqualToNumber("line2Accessible.selectionStart", 3);
            shouldBeEqualToNumber("line2Accessible.selectionEnd", 3);

            var contenteditable = document.getElementById('contenteditable-div');
            contenteditable.focus();
            var line1 = document.getElementById('paragraph1').firstChild;
            var line2 = document.getElementById('paragraph1').lastChild;
            var line3 = document.getElementById('paragraph2').firstChild;
            var contenteditableLines = [ line1, line2, line3 ];
            var contenteditableAccessible =
                accessibilityController.accessibleElementById(
                'contenteditable-div');

            for (var lineNumber = 0; lineNumber < 3; ++lineNumber) {
                var lineOffset = lineNumber * 7;
                // Paragraphs should have an extra blank line between them.
                if (lineNumber == 2)
                    lineOffset += 1;

                for (var characterOffset = 0; characterOffset < 7; ++characterOffset) {
                    caretPosition = lineOffset + characterOffset;

                    selectionRange.setStart(contenteditableLines[lineNumber], characterOffset);
                    selectionRange.setEnd(contenteditableLines[lineNumber], characterOffset);
                    selection.removeAllRanges();
                    selection.addRange(selectionRange);

                    shouldBeEqualToNumber("contenteditableAccessible.selectionStart",
                        caretPosition);
                    shouldBeEqualToNumber("contenteditableAccessible.selectionEnd",
                        caretPosition);
                    shouldBeEqualToNumber("contenteditableAccessible.selectionStartLineNumber",
                        lineNumber);
                    shouldBeEqualToNumber("contenteditableAccessible.selectionEndLineNumber",
                        lineNumber);
                }
            }

        }
    </script>

  </body>
</html>
