third_party/WebKit/LayoutTests/accessibility/contenteditable-caret-position-expected.txt

This tests that caret position is reported correctly for contenteditable elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textboxAccessible.selectionStart is 0
PASS textboxAccessible.selectionEnd is 0
PASS textboxAccessible.selectionStart is 1
PASS textboxAccessible.selectionEnd is 1
PASS textboxAccessible.selectionStart is 0
PASS textboxAccessible.selectionEnd is 0
PASS line2Accessible.selectionStart is 0
PASS line2Accessible.selectionEnd is 0
PASS line2Accessible.selectionStart is 3
PASS line2Accessible.selectionEnd is 3
PASS contenteditableAccessible.selectionStart is 0
PASS contenteditableAccessible.selectionEnd is 0
PASS contenteditableAccessible.selectionStartLineNumber is 0
PASS contenteditableAccessible.selectionEndLineNumber is 0
PASS contenteditableAccessible.selectionStart is 1
PASS contenteditableAccessible.selectionEnd is 1
PASS contenteditableAccessible.selectionStartLineNumber is 0
PASS contenteditableAccessible.selectionEndLineNumber is 0
PASS contenteditableAccessible.selectionStart is 2
PASS contenteditableAccessible.selectionEnd is 2
PASS contenteditableAccessible.selectionStartLineNumber is 0
PASS contenteditableAccessible.selectionEndLineNumber is 0
PASS contenteditableAccessible.selectionStart is 3
PASS contenteditableAccessible.selectionEnd is 3
PASS contenteditableAccessible.selectionStartLineNumber is 0
PASS contenteditableAccessible.selectionEndLineNumber is 0
PASS contenteditableAccessible.selectionStart is 4
PASS contenteditableAccessible.selectionEnd is 4
PASS contenteditableAccessible.selectionStartLineNumber is 0
PASS contenteditableAccessible.selectionEndLineNumber is 0
PASS contenteditableAccessible.selectionStart is 5
PASS contenteditableAccessible.selectionEnd is 5
PASS contenteditableAccessible.selectionStartLineNumber is 0
PASS contenteditableAccessible.selectionEndLineNumber is 0
PASS contenteditableAccessible.selectionStart is 6
PASS contenteditableAccessible.selectionEnd is 6
PASS contenteditableAccessible.selectionStartLineNumber is 0
PASS contenteditableAccessible.selectionEndLineNumber is 0
PASS contenteditableAccessible.selectionStart is 7
PASS contenteditableAccessible.selectionEnd is 7
PASS contenteditableAccessible.selectionStartLineNumber is 1
PASS contenteditableAccessible.selectionEndLineNumber is 1
PASS contenteditableAccessible.selectionStart is 8
PASS contenteditableAccessible.selectionEnd is 8
PASS contenteditableAccessible.selectionStartLineNumber is 1
PASS contenteditableAccessible.selectionEndLineNumber is 1
PASS contenteditableAccessible.selectionStart is 9
PASS contenteditableAccessible.selectionEnd is 9
PASS contenteditableAccessible.selectionStartLineNumber is 1
PASS contenteditableAccessible.selectionEndLineNumber is 1
PASS contenteditableAccessible.selectionStart is 10
PASS contenteditableAccessible.selectionEnd is 10
PASS contenteditableAccessible.selectionStartLineNumber is 1
PASS contenteditableAccessible.selectionEndLineNumber is 1
PASS contenteditableAccessible.selectionStart is 11
PASS contenteditableAccessible.selectionEnd is 11
PASS contenteditableAccessible.selectionStartLineNumber is 1
PASS contenteditableAccessible.selectionEndLineNumber is 1
PASS contenteditableAccessible.selectionStart is 12
PASS contenteditableAccessible.selectionEnd is 12
PASS contenteditableAccessible.selectionStartLineNumber is 1
PASS contenteditableAccessible.selectionEndLineNumber is 1
PASS contenteditableAccessible.selectionStart is 13
PASS contenteditableAccessible.selectionEnd is 13
PASS contenteditableAccessible.selectionStartLineNumber is 1
PASS contenteditableAccessible.selectionEndLineNumber is 1
PASS contenteditableAccessible.selectionStart is 15
PASS contenteditableAccessible.selectionEnd is 15
FAIL contenteditableAccessible.selectionStartLineNumber should be 2. Was 1.
FAIL contenteditableAccessible.selectionEndLineNumber should be 2. Was 1.
PASS contenteditableAccessible.selectionStart is 16
PASS contenteditableAccessible.selectionEnd is 16
PASS contenteditableAccessible.selectionStartLineNumber is 2
PASS contenteditableAccessible.selectionEndLineNumber is 2
PASS contenteditableAccessible.selectionStart is 17
PASS contenteditableAccessible.selectionEnd is 17
PASS contenteditableAccessible.selectionStartLineNumber is 2
PASS contenteditableAccessible.selectionEndLineNumber is 2
PASS contenteditableAccessible.selectionStart is 18
PASS contenteditableAccessible.selectionEnd is 18
PASS contenteditableAccessible.selectionStartLineNumber is 2
PASS contenteditableAccessible.selectionEndLineNumber is 2
PASS contenteditableAccessible.selectionStart is 19
PASS contenteditableAccessible.selectionEnd is 19
PASS contenteditableAccessible.selectionStartLineNumber is 2
PASS contenteditableAccessible.selectionEndLineNumber is 2
PASS contenteditableAccessible.selectionStart is 20
PASS contenteditableAccessible.selectionEnd is 20
PASS contenteditableAccessible.selectionStartLineNumber is 2
PASS contenteditableAccessible.selectionEndLineNumber is 2
PASS contenteditableAccessible.selectionStart is 21
PASS contenteditableAccessible.selectionEnd is 21
PASS contenteditableAccessible.selectionStartLineNumber is 2
PASS contenteditableAccessible.selectionEndLineNumber is 2
PASS successfullyParsed is true

TEST COMPLETE
Line 1

Line 1
Line 2

Line 3
